fix code

+++ src/main/java/com/tiximax/txm/Service/PartialShipmentService.java
@@ -613,7 +613,7 @@ public class PartialShipmentService {
                 .stream()
                 .collect(Collectors.toMap(
                         row -> (String) row[0],
-                        row -> (BigDecimal) row[1]
+                        row -> BigDecimal.valueOf((Double) row[1])
                 ));
 
             BigDecimal collectWeight =
